requirements.txt
huggingface-hub
streamlit
scrapegraph-py
pdfkit
agno
openai
duckduckgo-search
yfinance
google-search-results
PyPDF2
python-docx
openpyxl
pandas
Pillow
mcp
python-dotenv
openai
-e .